vstream: move functionality to vstreamManager
We need more elaborate data structures for handling journal events.
